fixs and utils
